requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
aldjemy==2.6
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cssselect2==0.7.0
CurrencyConverter==0.17.32
defusedxml==0.7.1
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.1.0
num2words==0.5.13
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pdfkit==1.0.0
phonenumbers==8.13.51
pika==1.3.2
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyPDF2==3.0.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
tinycss2==1.4.0
twilio==9.3.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.7.0
XlsxWriter==3.2.0
yarl==1.14.0
